refactor: login flow adapted
