+++ frontend/src/pages/login/index.page.tsx
@@ -1,1 +1,1 @@
-export { default } from './LoginPage'
+export { default } from './Login';
